Update PlayerCharacter
Add a pointer to the focused actor.
Add a boolean bPlayerCanMove.
Movement and ControlRotation allowed only if PlayerCanMove = true.
Add 2 inputs Interact and Cancel bound to 2 new functions.
Interact calls the Interact() function on the focused actor if it implements the InteractInterface.
Cancel sets unlocks player movement and removes focused widget from the screen.
Refoactor ReachInFront to return a bool.
Keep a reference to custom PlayerController to make things easier.