Update PlayerCharacter
Add a Camera and a SpringArm. Spring arm is used to grab an item.
Add input binding to grab and drop items.
Add input binding to quit the game.
Add a PlayerCanInteract boolean and add get/set functions.
Cleanup useless functions related to HUD. HUD is now handled in the widgets.